finish CategoriesScreen

+++ user/src/main/java/org/the_chance/honeymart/ui/feature/category/CategoryItem.kt
@@ -30,7 +30,7 @@ import org.the_chance.honymart.ui.theme.white100
  * Created by Aziza Helmy on 7/27/2023.
  */
 @Composable
-fun CategoryItem(state: CategoryUiState, onCategoryClicked: () -> Unit) {
+fun CategoryItem(state: CategoryUiState, onCategoryClicked: (Long) -> Unit) {
     Column(
         horizontalAlignment = CenterHorizontally,
         verticalArrangement = Arrangement.Center,
